Allow specifying a custom config file.
Augments the Config class to accept an optional filename parameter to
specify a custom config file. Adds an argument to the CLI to pass the config
file through to the Config class.

Adds tests to ensure existing functionality is maintained, and paths are
expanded to abspaths properly.  Adds a class variable "persist" to the Config
class to prevent writing files to disk during testing.